pyproject.toml
[tool.poetry]
name = "metacrawler"
version = "0.1.0"
description = ""
authors = ["Shannon Gray <shanngray@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
neo4j = "^5.21.0"
langchain = "^0.2.5"
langchain-community = "^0.2.5"
python-dotenv = "^1.0.1"
langchain-groq = "^0.1.5"
python-magic = "^0.4.27"
tiktoken = "^0.7.0"
langchain-cohere = "^0.1.8"
mockfs = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
